# backend/src/api/retailer/dto/responses/retailer-profile.response.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { Role } from 'prisma/generated/prisma/client';

export class RetailerProfileListResponseDto {
  @ApiProperty()
  id!: string;

  @ApiProperty()
  email!: string;

  @ApiProperty()
  username!: string;

  @ApiProperty({ enum: Role })
  role!: Role;

  @ApiProperty()
  companyName!: string;

  @ApiProperty()
  businessAddress!: string;
}