# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/verify/{batchId}": {
      "get": {
        "operationId": "VerifyController_verifyBatch",
        "parameters": [
          {
            "name": "batchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produce batch verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/VerifyProduceResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Verify"
        ]
      }
    },
    "/retailer/batches": {
      "get": {
        "operationId": "RetailerController_listAssignedBatches",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter produce batches by status",
            "schema": {
              "example": "ARRIVED",
              "type": "string",
              "enum": [
                "DRAFT",
                "PENDING_CHAIN",
                "ONCHAIN_CONFIRMED",
                "IN_TRANSIT",
                "ARRIVED",
                "RETAILER_VERIFIED",
                "ARCHIVED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by produce name, farm name, or batch ID (case-insensitive)",
            "schema": {
              "example": "mango",
              "type": "string"
            }
          },
          {
            "name": "harvestFrom",
            "required": false,
            "in": "query",
            "description": "Filter batches harvested on or after this date (ISO string)",
            "schema": {
              "example": "2024-01-01",
              "type": "string"
            }
          },
          {
            "name": "harvestTo",
            "required": false,
            "in": "query",
            "description": "Filter batches harvested on or before this date (ISO string)",
            "schema": {
              "example": "2024-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assigned produce batches retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ProduceListResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Retailer"
        ]
      }
    },
    "/retailer/verify/{batchId}": {
      "post": {
        "operationId": "RetailerController_verifyBatch",
        "parameters": [
          {
            "name": "batchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Retailer"
        ]
      }
    },
    "/retailer/profiles": {
      "get": {
        "operationId": "RetailerController_listRetailerProfiles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Retailers with profiles retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/RetailerProfileListResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Retailer"
        ]
      }
    },
...
    "/reports": {
      "post": {
        "operationId": "ReportController_createReport",
        "parameters": [
          {
            "name": "farmId",
            "required": false,
            "in": "query",
            "description": "Filter by farm ID (used for farm/produce related reports)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "description": "Filter farms by state",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "district",
            "required": false,
            "in": "query",
            "description": "Filter farms by district",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Filter data on or after this date (ISO string)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "Filter data on or before this date (ISO string)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Generic status filter (interpreted per report type, e.g. subsidy/produce/program status)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "description": "Action filter (used for activity reports)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ethToMyr",
            "required": false,
            "in": "query",
            "description": "ETH to MYR exchange rate at the time of report generation (used for subsidy reports)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minFarmSize",
            "required": false,
            "in": "query",
            "description": "Minimum farm size (inclusive, numeric string)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxFarmSize",
            "required": false,
            "in": "query",
            "description": "Maximum farm size (inclusive, numeric string)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "farmVerificationStatus",
            "required": false,
            "in": "query",
            "description": "Filter farms by verification status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "VERIFIED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "programType",
            "required": false,
            "in": "query",
            "description": "Filter programs by type",
            "schema": {
              "type": "string",
              "enum": [
                "drought",
                "flood",
                "crop_loss",
                "manual"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Report created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ReportResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Report"
        ]
      },
      "get": {
        "operationId": "ReportController_listReports",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "reportType",
            "required": false,
            "in": "query",
            "description": "Filter by report type",
            "schema": {
              "type": "string",
              "enum": [
                "FARM_SUMMARY",
                "SUBSIDY_REPORT",
                "PRODUCE_REPORT",
                "PROGRAM_REPORT",
                "FINANCIAL_REPORT",
                "ACTIVITY_REPORT",
                "CUSTOM"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by report status",
            "schema": {
              "type": "string",
              "enum": [
                "GENERATING",
                "COMPLETED",
                "FAILED"
              ]
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Filter reports created on or after this date (ISO string)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "Filter reports created on or before this date (ISO string)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reports retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ReportResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Report"
        ]
      }
    },
    "/reports/{id}": {
      "get": {
        "operationId": "ReportController_getReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ReportResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Report"
        ]
      }
    },
    "/reports/{id}/download": {
      "get": {
        "operationId": "ReportController_downloadReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Report"
        ]
      }
    }
  },
  "info": {
    "title": "My API",
    "description": "The backend API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CommonResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "FarmSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "description": "Average rating for the farm",
            "example": 4.5
          },
          "ratingCount": {
            "type": "number",
            "description": "Number of reviews contributing to the rating",
            "example": 12
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "state",
          "district",
          "rating",
          "ratingCount"
        ]
      },
      "ProduceCertificateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "produceId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "Certificate label/type (e.g., Organic, Halal)."
          },
          "ipfsUrl": {
            "type": "string",
            "description": "IPFS gateway URL pointing to the uploaded certificate."
          },
          "verifiedOnChain": {
            "type": "boolean",
            "description": "Whether this certificate is verified on-chain"
          },
          "issuedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the certificate was issued (if provided).",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "description": "Arbitrary metadata stored with the certificate",
            "nullable": true,
            "additionalProperties": true
          },
          "fileName": {
            "type": "string",
            "description": "Original file name as uploaded",
            "nullable": true
          },
          "mimeType": {
            "type": "string",
            "description": "MIME type of the uploaded file",
            "nullable": true
          },
          "fileSize": {
            "type": "number",
            "description": "Size of the uploaded file in bytes",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the certificate was created"
          }
        },
        "required": [
          "id",
          "produceId",
          "type",
          "ipfsUrl",
          "verifiedOnChain",
          "createdAt"
        ]
      },
      "QRCodeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique ID of the QR code"
          },
          "produceId": {
            "type": "string",
            "description": "ID of the associated produce"
          },
          "verifyUrl": {
            "type": "string",
            "description": "URL used to generate the QR code"
          },
          "imageUrl": {
            "type": "string",
            "description": "Image URL of the QR code"
          },
          "qrHash": {
            "type": "string",
            "description": "Optional QR hash for verification",
            "nullable": true
          },
          "isPublicQR": {
            "type": "boolean",
            "description": "Indicates if the QR is publicly shareable"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          }
        },
        "required": [
          "id",
          "produceId",
          "verifyUrl",
          "imageUrl",
          "isPublicQR",
          "createdAt"
        ]
      },
      "ProduceListResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "farmId": {
            "type": "string"
          },
          "farm": {
            "description": "Associated farm summary",
            "allOf": [
              {
                "$ref": "#/components/schemas/FarmSummaryDto"
              }
            ]
          },
          "category": {
            "type": "string"
          },
          "batchId": {
            "type": "string"
          },
          "certifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProduceCertificateDto"
            }
          },
          "harvestDate": {
            "format": "date-time",
            "type": "string"
          },
          "blockchainTx": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "description": "Current lifecycle status of the produce batch",
            "enum": [
              "DRAFT",
              "PENDING_CHAIN",
              "ONCHAIN_CONFIRMED",
              "IN_TRANSIT",
              "ARRIVED",
              "RETAILER_VERIFIED",
              "ARCHIVED"
            ],
            "default": "DRAFT"
          },
          "retailerId": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "number",
            "description": "Recorded quantity of the produce batch",
            "example": 150
          },
          "unit": {
            "type": "string",
            "description": "Unit used for the recorded quantity",
            "enum": [
              "KG",
              "G",
              "TONNE",
              "PCS",
              "BUNCH",
              "TRAY",
              "L",
              "ML"
            ],
            "default": "KG"
          },
          "imageUrl": {
            "type": "string",
            "description": "Image URL for the produce batch",
            "nullable": true
          },
          "qrCode": {
            "description": "Data URI for the QR code",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/QRCodeDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "farmId",
          "category",
          "batchId",
          "certifications",
          "harvestDate",
          "blockchainTx",
          "status",
          "retailerId",
          "unit"
        ]
      },
...
      "ProduceFarmInfoDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Organic Tomato",
            "description": "Name of the produce batch"
          },
          "harvestDate": {
            "format": "date-time",
            "type": "string",
            "example": "2025-10-30T00:00:00.000Z",
            "description": "Harvest date of the produce batch in ISO format"
          },
          "certifications": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "farm": {
            "type": "string",
            "example": "Ali’s Organic Farm",
            "nullable": true,
            "description": "Origin farm name"
          }
        },
        "required": [
          "name",
          "harvestDate",
          "farm"
        ]
      },
      "BlockchainProofDto": {
        "type": "object",
        "properties": {
          "onChainHash": {
            "type": "string",
            "example": "0xabc123...def456",
            "description": "Produce hash stored on-chain for immutability verification"
          },
          "offChainHash": {
            "type": "string",
            "example": "0xabc123...def456",
            "description": "Locally computed produce hash for integrity comparison"
          },
          "blockchainTx": {
            "type": "string",
            "example": "0x10e12b152edb6679959b095faf6918e8c2cae6121ef91eab57d4302eea273325",
            "description": "Blockchain transaction hash where the produce was recorded",
            "nullable": true
          }
        },
        "required": [
          "onChainHash",
          "offChainHash"
        ]
      },
      "VerifyProduceResponseDto": {
        "type": "object",
        "properties": {
          "batchId": {
            "type": "string",
            "example": "BATCH123",
            "description": "Unique identifier of the produce batch"
          },
          "status": {
            "type": "string",
            "example": "ONCHAIN_CONFIRMED",
            "description": "Current status of the produce record in the system"
          },
          "isPublicQR": {
            "type": "boolean",
            "example": true,
            "description": "Indicates if the QR code for this produce is publicly verifiable"
          },
          "produce": {
            "description": "Off-chain produce details used for verification display",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProduceFarmInfoDto"
              }
            ]
          },
          "blockchain": {
            "description": "Blockchain verification and transaction details",
            "allOf": [
              {
                "$ref": "#/components/schemas/BlockchainProofDto"
              }
            ]
          }
        },
        "required": [
          "batchId",
          "status",
          "isPublicQR",
          "produce",
          "blockchain"
        ]
      },
      "RetailerProfileListResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "FARMER",
              "RETAILER",
              "GOVERNMENT_AGENCY",
              "ADMIN"
            ]
          },
          "companyName": {
            "type": "string"
          },
          "businessAddress": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "role",
          "companyName",
          "businessAddress"
        ]
      },
...
      "ReportResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "reportType": {
            "type": "string",
            "enum": [
              "FARM_SUMMARY",
              "SUBSIDY_REPORT",
              "PRODUCE_REPORT",
              "PROGRAM_REPORT",
              "FINANCIAL_REPORT",
              "ACTIVITY_REPORT",
              "CUSTOM"
            ]
          },
          "title": {
            "type": "string"
          },
          "parameters": {
            "type": "object"
          },
          "fileUrl": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "GENERATING",
              "COMPLETED",
              "FAILED"
            ]
          },
          "errorMessage": {
            "type": "string"
          },
          "generatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "reportType",
          "title",
          "status",
          "createdAt"
        ]
      },
      "CreateReportDto": {
        "type": "object",
        "properties": {
          "reportType": {
            "type": "string",
            "enum": [
              "FARM_SUMMARY",
              "SUBSIDY_REPORT",
              "PRODUCE_REPORT",
              "PROGRAM_REPORT",
              "FINANCIAL_REPORT",
              "ACTIVITY_REPORT",
              "CUSTOM"
            ],
            "description": "Type of report to generate"
          },
          "title": {
            "type": "string",
            "description": "Report title"
          }
        },
        "required": [
          "reportType",
          "title"
        ]
      }
    }
  }
}